@@ -3,6 +3,7 @@ package BGU.Group13B.backend.Repositories.Implementations.PurchaseHistoryReposit
 import BGU.Group13B.backend.Repositories.Interfaces.IPurchaseHistoryRepository;
 import BGU.Group13B.backend.User.BasketProduct;
 import BGU.Group13B.backend.User.PurchaseHistory;
+import BGU.Group13B.service.Response;
 
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -28,13 +29,14 @@ public class PurchaseHistoryRepositoryAsList implements IPurchaseHistoryReposito
         return purchaseHistories.stream().anyMatch(purchaseHistory -> purchaseHistory.getUserId() == userId && purchaseHistory.getStoreId() == storeId);
     }
     @Override
-    public String getAllPurchases(int userId) {
-        List<PurchaseHistory> purchases = getAllPurchasesAsList(userId);
-        StringBuilder sb = new StringBuilder();
-        for (PurchaseHistory purchaseHistory : purchases) {
-            sb.append(purchaseHistory.toString()).append("\n");
+    public List<PurchaseHistory> getAllPurchases(int userId) {
+        List<PurchaseHistory> purchases = new LinkedList<>();
+        for (PurchaseHistory purchaseHistory : purchaseHistories) {
+            if (purchaseHistory.getUserId() == userId) {
+                purchases.add(purchaseHistory);
+            }
         }
-        return sb.toString();
+        return purchases;
     }
     @Override
     public PurchaseHistory addPurchase(int userId, int storeId, ConcurrentLinkedQueue<BasketProduct> products, double price) {
@@ -48,15 +50,6 @@ public class PurchaseHistoryRepositoryAsList implements IPurchaseHistoryReposito
         return purchaseHistory;
     }
 
-    private List<PurchaseHistory> getAllPurchasesAsList(int userId) {
-        List<PurchaseHistory> purchases = new LinkedList<>();
-        for (PurchaseHistory purchaseHistory : purchaseHistories) {
-            if (purchaseHistory.getUserId() == userId) {
-                purchases.add(purchaseHistory);
-            }
-        }
-        return purchases;
-    }
 
     @Override
     public void reset() {
