@@ -2,21 +2,26 @@ package BGU.Group13B.backend.User;
 
 import BGU.Group13B.backend.Pair;
 import BGU.Group13B.backend.storePackage.Product;
-import BGU.Group13B.service.BroadCaster;
-import BGU.Group13B.service.PushNotification;
+import BGU.Group13B.service.*;
 import BGU.Group13B.service.entity.ServiceBasketProduct;
 import org.mindrot.jbcrypt.BCrypt;
 import BGU.Group13B.backend.Repositories.Interfaces.IMessageRepository;
 import BGU.Group13B.backend.Repositories.Interfaces.IPurchaseHistoryRepository;
 import BGU.Group13B.backend.Repositories.Interfaces.IUserPermissionRepository;
 import BGU.Group13B.backend.storePackage.Market;
+import BGU.Group13B.backend.storePackage.Product;
 import BGU.Group13B.backend.storePackage.Review;
 import BGU.Group13B.backend.storePackage.permissions.NoPermissionException;
-import BGU.Group13B.service.SingletonCollection;
+import BGU.Group13B.service.BroadCaster;
 //eyal import
 import java.time.LocalDate;
+
+import org.mindrot.jbcrypt.BCrypt;
+
+
 import java.util.HashMap;
 import java.util.List;
+import java.util.Set;
 import java.util.regex.Pattern;
 
 public class User {
@@ -73,8 +78,8 @@ public class User {
         this.answer3 = "";
         this.messageId = 1;
         this.isLoggedIn = false;
-        this.messageNotification= false;
-        this.reviewedStoreNotification= false;
+        this.messageNotification = false;
+        this.reviewedStoreNotification = false;
     }
 
 
@@ -95,17 +100,20 @@ public class User {
     //#15
     //returns User on success (for future functionalities)
     public User register(String userName, String password, String email, String answer1, String answer2, String answer3, LocalDate birthdate) {
-        checkRegisterInfo(userName, password, email);
+        checkRegisterInfo(userName, password, email, birthdate);
         //updates the user info upon registration - no longer a guest
-        updateUserDetail(userName, password, email, answer1, answer2, answer3,birthdate);
+        updateUserDetail(userName, password, email, answer1, answer2, answer3, birthdate);
         this.userPermissions.register(this.userId);
         return this;
     }
 
-    private void checkRegisterInfo(String userName, String password, String email) {
+    private void checkRegisterInfo(String userName, String password, String email, LocalDate birthdate) {
         String usernameRegex = "^[a-zA-Z0-9_-]{4,16}$"; // 4-16 characters, letters/numbers/underscore/hyphen
         String passwordRegex = "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)[A-Za-z\\d]{8,}$"; // need at least 8 characters, 1 uppercase, 1 lowercase, 1 number)
         String emailRegex = "^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$"; // checks email validation
+        if (birthdate == null) {
+            throw new IllegalArgumentException("enter birthdate bro");
+        }
         if (!Pattern.matches(usernameRegex, userName)) {
             throw new IllegalArgumentException("Invalid username. Username must be 4-16 characters long and can only contain letters, numbers, underscores, or hyphens.");
         }
@@ -152,17 +160,15 @@ public class User {
 
 
     }
-    public void fetchMessages() throws NoPermissionException
-    {
+
+    public void fetchMessages() throws NoPermissionException {
         if (!isRegistered())
             throw new NoPermissionException("Only registered users can fetch messages");
-        if(getAndSetMessageNotification())
-        {
-            BroadCaster.broadcast(this.userId,"New Message");
+        if (getAndSetMessageNotification()) {
+            BroadCaster.broadcast(this.userId, "New Message");
         }
-        if(getAndSetReviewedStoreNotification())
-        {
-            BroadCaster.broadcast(this.userId,"New Review");
+        if (getAndSetReviewedStoreNotification()) {
+            BroadCaster.broadcast(this.userId, "New Review");
         }
     }
 
@@ -201,12 +207,12 @@ public class User {
         if (!isAdmin())
             throw new NoPermissionException("Only admin can send massages");
         messageRepository.sendMassage(Message.constractMessage(this.userName, getAndIncrementMessageId(), header, massage, receiverId));
-        User receiverNext=SingletonCollection.getUserRepository().getUserByUsername(receiverId);
+        User receiverNext = SingletonCollection.getUserRepository().getUserByUsername(receiverId);
         //if(!PushNotification.pushNotification("New Message",receiverNext.getUserId()))
-         if(!receiverNext.isLoggedIn() || !BroadCaster.broadcast(receiverNext.userId,"New Message"))
+        if (!receiverNext.isLoggedIn() || !BroadCaster.broadcast(receiverNext.userId, "New Message"))
             receiverNext.setMessageNotification(true);
-        System.out.println("ReceiverNext: "+receiverNext.userName+" LoggedIn: "+receiverNext.isLoggedIn());
-        System.out.println("Status: "+receiverNext.getMessageNotification());
+        System.out.println("ReceiverNext: " + receiverNext.userName + " LoggedIn: " + receiverNext.isLoggedIn());
+        System.out.println("Status: " + receiverNext.getMessageNotification());
     }
 
     //tdsfds
@@ -218,15 +224,16 @@ public class User {
             throw new IllegalArgumentException("no complaint to answer");
         messageRepository.markAsRead(regularMessageToReply.getReceiverId(), regularMessageToReply.getSenderId(), regularMessageToReply.getMessageId());
         messageRepository.sendMassage(Message.constractMessage(this.userName, getAndIncrementMessageId(), "RE: " + regularMessageToReply.getHeader(), answer, regularMessageToReply.getSenderId()));
-        User receiverNext=SingletonCollection.getUserRepository().getUserByUsername(regularMessageToReply.getSenderId());
-        if(!receiverNext.isLoggedIn() || !BroadCaster.broadcast(receiverNext.userId,"New Message"))
+        User receiverNext = SingletonCollection.getUserRepository().getUserByUsername(regularMessageToReply.getSenderId());
+        if (!receiverNext.isLoggedIn() || !BroadCaster.broadcast(receiverNext.userId, "New Message"))
             receiverNext.setMessageNotification(true);
         regularMessageToReply = null;
     }
-    public void clearMessageToReply()
-    {
-        regularMessageToReply =null;
+
+    public void clearMessageToReply() {
+        regularMessageToReply = null;
     }
+
     public Message readMassage() throws NoPermissionException {
         if (!isRegistered())
             throw new NoPermissionException("Only registered users can read massages");
@@ -243,19 +250,19 @@ public class User {
         if (regularMessageToReply == null)
             throw new IllegalArgumentException("no message to answer");
         messageRepository.sendMassage(Message.constractMessage(this.userName, getAndIncrementMessageId(), "RE: " + regularMessageToReply.getHeader(), answer, regularMessageToReply.getSenderId()));
-        User receiverNext=SingletonCollection.getUserRepository().getUserByUsername(regularMessageToReply.getSenderId());
+        User receiverNext = SingletonCollection.getUserRepository().getUserByUsername(regularMessageToReply.getSenderId());
 
-        if(!receiverNext.isLoggedIn() || !BroadCaster.broadcast(receiverNext.userId,"New Message"))
+        if (!receiverNext.isLoggedIn() || !BroadCaster.broadcast(receiverNext.userId, "New Message"))
             receiverNext.setMessageNotification(true);
         regularMessageToReply = null;
     }
 
     public void sendMassageBroad(String receiverName, String header, String massage) throws NoPermissionException {
         messageRepository.sendMassage(Message.constractMessage(this.userName, getAndIncrementMessageId(), header, massage, receiverName));
-        User receiverNext=SingletonCollection.getUserRepository().getUserByUsername(receiverName);
-        if(!BroadCaster.broadcast(receiverNext.userId,"New Message"))
+        User receiverNext = SingletonCollection.getUserRepository().getUserByUsername(receiverName);
+        if (!BroadCaster.broadcast(receiverNext.userId, "New Message"))
             receiverNext.setMessageNotification(true);
-        System.out.println("Status: "+receiverNext.getMessageNotification());
+        System.out.println("Status: " + receiverNext.getMessageNotification());
     }
 
     public Message readOldMessage() throws NoPermissionException {
@@ -303,8 +310,8 @@ public class User {
         assert regularMessageToReply.getReceiverId().matches("-?\\d+");
         market.markAsCompleted(regularMessageToReply.getSenderId(), regularMessageToReply.getMessageId(), this.userId, Integer.parseInt(regularMessageToReply.getReceiverId()));
         messageRepository.sendMassage(Message.constractMessage(this.userName, getAndIncrementMessageId(), "RE: " + regularMessageToReply.getHeader(), answer, regularMessageToReply.getSenderId()));
-        User receiverNext=SingletonCollection.getUserRepository().getUserByUsername(regularMessageToReply.getSenderId());
-        if(!BroadCaster.broadcast(receiverNext.userId,"New Message"))
+        User receiverNext = SingletonCollection.getUserRepository().getUserByUsername(regularMessageToReply.getSenderId());
+        if (!BroadCaster.broadcast(receiverNext.userId, "New Message"))
             receiverNext.setMessageNotification(true);
         regularMessageToReply = null;
     }
@@ -336,15 +343,17 @@ public class User {
     public Review getReview(int storeId, int productId) throws NoPermissionException {
         return market.getReview(storeId, productId, this.userId);
     }
-    public List<Review> getAllReviews(int storeId, int productId,int userId) throws NoPermissionException {
-        return market.getAllReviews(storeId, productId);
+
+    public List<Review> getAllReviews(int storeId, int productId, int userId) throws NoPermissionException {
+        return market.getAllReviews(productId, storeId);
     }
 
     //#26
     public float getProductScore(int storeId, int productId) throws NoPermissionException {
         return market.getProductScore(storeId, productId, userId);
     }
-    public float getProductScoreUser(int userIdTarget,int storeId, int productId) throws NoPermissionException {
+
+    public float getProductScoreUser(int userIdTarget, int storeId, int productId) throws NoPermissionException {
         return market.getProductScoreUser(storeId, productId, userIdTarget);
     }
 
@@ -502,8 +511,8 @@ public class User {
     }
 
 
-    public void addToCart(int storeId, int productId) {
-        cart.addProductToCart(storeId, productId);
+    public void addToCart(int storeId, int productId, int amount, double newPrice) {
+        cart.addProductToCart(storeId, productId, amount, newPrice);
     }
 
     public void addPermission(int storeId, UserPermissions.StoreRole storeRole) {
@@ -516,6 +525,22 @@ public class User {
 
     public UserPermissions getUserPermissions() {
         return userPermissions;
+    }
+
+    public void addIndividualPermission(int storeId, UserPermissions.IndividualPermission individualPermission) {
+        userPermissions.addIndividualPermission(storeId, individualPermission);
+    }
+
+    public void deleteIndividualPermission(int storeId, UserPermissions.IndividualPermission individualPermission) {
+        userPermissions.deleteIndividualPermission(storeId, individualPermission);
+    }
+
+    public Set<UserPermissions.IndividualPermission> getIndividualPermissions(int storeId) {
+        return userPermissions.getIndividualPermissions(storeId);
+    }
+
+    public void removeAllIndividualPermissions(int storeId) {
+        userPermissions.removeAllIndividualPermissions(storeId);
     }
 
 
@@ -531,11 +556,13 @@ public class User {
     public synchronized boolean getMessageNotification() {
         return messageNotification;
     }
+
     public synchronized boolean getAndSetMessageNotification() {
         boolean messageNotification = this.messageNotification;
-        this.messageNotification= false;
+        this.messageNotification = false;
         return messageNotification;
     }
+
     public synchronized void setMessageNotification(boolean notifications) {
         this.messageNotification = notifications;
     }
@@ -543,11 +570,13 @@ public class User {
     public synchronized boolean getReviewedStoreNotification() {
         return reviewedStoreNotification;
     }
+
     public synchronized boolean getAndSetReviewedStoreNotification() {
         boolean reviewedStoreNotification = this.reviewedStoreNotification;
-        this.reviewedStoreNotification= false;
+        this.reviewedStoreNotification = false;
         return reviewedStoreNotification;
     }
+
     public synchronized void setReviewedStoreNotification(boolean notifications) {
         this.reviewedStoreNotification = notifications;
     }
@@ -565,11 +594,23 @@ public class User {
         cart.cancelPurchase();
     }
 
-    public String getPurchaseHistory() {
-       return purchaseHistoryRepository.getAllPurchases(userId);
+    public List<PurchaseHistory> getPurchaseHistory() {
+        return purchaseHistoryRepository.getAllPurchases(userId);
     }
 
     public LocalDate getDateOfBirth() {
         return dateOfBirth;
+    }
+
+    public UserCard getUserCard(){
+        List<Pair<Integer, String>> pairs = userPermissions.getStoresAndRoles();
+        StringBuilder stringBuilder = new StringBuilder();
+        for(Pair<Integer, String> pair : pairs){
+            stringBuilder.append("Store: ");
+            stringBuilder.append(pair.getFirst());
+            stringBuilder.append(" Role: ");
+            stringBuilder.append(pair.getSecond());
+        }
+        return new UserCard(userId, userName, email, stringBuilder.toString());
     }
 }
