@@ -6,6 +6,7 @@ import BGU.Group13B.backend.Repositories.Interfaces.*;
 import BGU.Group13B.backend.User.*;
 import BGU.Group13B.backend.storePackage.delivery.DeliveryAdapter;
 import BGU.Group13B.backend.storePackage.newDiscoutns.PurchasePolicy;
+import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.Condition;
 import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.DiscountPolicy;
 import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.StoreDiscount;
 import BGU.Group13B.backend.storePackage.payment.PaymentAdapter;
@@ -14,6 +15,7 @@ import BGU.Group13B.backend.storePackage.purchaseBounders.PurchaseExceedsPolicyE
 import BGU.Group13B.service.BroadCaster;
 import BGU.Group13B.service.SingletonCollection;
 import BGU.Group13B.service.callbacks.AddToUserCart;
+import BGU.Group13B.service.info.DiscountAccumulationTreeInfo;
 import BGU.Group13B.service.info.StoreInfo;
 
 
@@ -107,6 +109,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StaffPermission
     public void addOwner(int userId, int newOwnerId) throws NoPermissionException, ChangePermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add an owner to store " + this.storeId);
@@ -116,11 +119,12 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StaffPermission
     public void removeOwner(int userId, int removeOwnerId) throws NoPermissionException, ChangePermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to remove an owner to store " + this.storeId);
-        List<Integer> removeUsersList = storePermission.removeOwnerPermission(removeOwnerId, userId, false);
-        for (Integer removeUserId : removeUsersList) {
+        Set<Integer> removeUsersList = storePermission.removeOwnerPermission(removeOwnerId, userId, false);
+        for(Integer removeUserId: removeUsersList){
             userRepository.getUser(removeUserId).deletePermission(storeId);
         }
 
@@ -128,6 +132,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StaffPermission
     public void addManager(int userId, int newManagerId) throws NoPermissionException, ChangePermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add a manager to store " + this.storeId);
@@ -137,6 +142,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StaffPermission
     public void removeManager(int userId, int removeManagerId) throws NoPermissionException, ChangePermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to remove a manager to store " + this.storeId);
@@ -146,6 +152,27 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StaffPermission
+    public void addPermissionToManager(int userId, int managerId, UserPermissions.IndividualPermission individualPermission) throws NoPermissionException, ChangePermissionException {
+        if (!this.storePermission.checkPermission(userId, hidden))
+            throw new NoPermissionException("User " + userId + " has no permission to add permissions to managers in store " + this.storeId);
+        storePermission.addIndividualPermission(managerId, individualPermission);
+        userRepository.getUser(managerId).addIndividualPermission(storeId, individualPermission);
+    }
+
+    @DefaultFounderFunctionality
+    @DefaultOwnerFunctionality
+    @StaffPermission
+    public void removePermissionFromManager(int userId, int managerId, UserPermissions.IndividualPermission individualPermission) throws NoPermissionException, ChangePermissionException {
+        if (!this.storePermission.checkPermission(userId, hidden))
+            throw new NoPermissionException("User " + userId + " has no permission to remove permissions from managers in store " + this.storeId);
+        storePermission.removeIndividualPermission(managerId, individualPermission);
+        userRepository.getUser(managerId).deleteIndividualPermission(storeId, individualPermission);
+    }
+
+    @DefaultFounderFunctionality
+    @DefaultOwnerFunctionality
+    @WorkersInfoPermission
     public List<WorkerCard> getStoreWorkersInfo(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to get info about the store workers " + this.storeId);
@@ -160,6 +187,10 @@ public class Store {
         return storePermission.getStoreOwners();
     }
 
+    public StoreInfo getGemeralStoreInfo()  {
+        return new StoreInfo(this);
+    }
+
     public StorePermission getStorePermission() {
         return storePermission;
     }
@@ -171,6 +202,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @MessagesPermission
     public Message getUnreadMessages(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to read message of store " + this.storeId);
@@ -180,6 +212,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @MessagesPermission
     public Message getReadMessages(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to read message of store " + this.storeId);
@@ -188,6 +221,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @MessagesPermission
     public void markAsCompleted(String senderId, int messageId, int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to mark message as complete of store: " + this.storeId);
@@ -196,6 +230,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @MessagesPermission
     public void refreshMessages(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to handle message of store " + this.storeId);
@@ -303,11 +338,10 @@ public class Store {
     }
 
 
-    /*WARNING:default manager functionality is only 4.12: "answer store questions & appeals feature"
-       and 4.13: "get store purchase history feature"  **/
+
     @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public synchronized int addProduct(int userId, String productName, String category, double price, int stockQuantity, String description) throws NoPermissionException {
         /*
          * check if the user has permission to add product
@@ -342,16 +376,12 @@ public class Store {
                 Thread.currentThread().getStackTrace()[2].getMethodName());
     }
 
-    @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
-    @DefaultOwnerFunctionality
+
     public void purchaseProposalSubmit(int userId, int productId, double proposedPrice, int amount) throws NoPermissionException {
         /*
          * check if the user has permission to purchase proposal
          * */
-        if (!this.storePermission.checkPermission(userId, hidden))//the user should be loggedIn with permissions
-            throw new NoPermissionException("User " + userId + " has no permission to add product to store " + this.storeId);
-
+        this.storePermission.validateStoreVisibility(userId, hidden);
         if (amount <= 0)
             throw new IllegalArgumentException("Amount must be positive");
         if (proposedPrice <= 0)
@@ -362,15 +392,22 @@ public class Store {
         /*
          * Send alert to all the permitted Managers
          * */
-        Set<Integer> userIds = this.storePermission.getAllUsersWithPermission(Store.getCurrentMethodName());
+        //Set<Integer> userIds = this.storePermission.getAllUsersWithPermission(Store.getCurrentMethodName());
+        List<Integer> userIds = this.storePermission.getAllUsersWithIndividualPermission(UserPermissions.IndividualPermission.AUCTION);
         for (Integer id : userIds) {//wait for the interface in AlertManager.java to finish
-            alertManager.sendAlert(id, "User " + userId + " has submitted a purchase proposal for product " + productId + " in store " + this.storeId);
-            //fixme!!!!!!!!!
+            //alertManager.sendAlert(id, "User " + userId + " has submitted a purchase proposal for product " + productId + " in store " + this.storeId);
+            double originalPrice = productRepository.getProductById(productId).getPrice();
+            BroadCaster.broadcast(id,
+                    "BID["+storeId+","+productId+"]User has submitted a purchase proposal for product " + productId +
+                    "\n In store " + this.storeName +
+                            "\nOriginal price: " + originalPrice +
+                            "\nProposed price: " + proposedPrice +
+                            "\nAmount: " + amount + "\n");
         }
     }
 
     @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
+    @AuctionPermission
     @DefaultOwnerFunctionality
     public void purchaseProposalApprove(int managerId, int bidId) throws NoPermissionException {
         /*
@@ -382,18 +419,24 @@ public class Store {
                 orElseThrow(() -> new IllegalArgumentException("There is no such bid for store " + this.storeId));
 
         if (currentBid.isRejected())//fixme, use BroadCaster.broadcast
-            alertManager.sendAlert(managerId, "The bid for product " + currentBid.getProductId() + " in store " + this.storeId + " has been rejected already");
+            BroadCaster.broadcast(managerId,
+                    "The bid for product " + productRepository.getProductById(currentBid.getProductId()).getName()
+                            + " has been rejected already");
+            //alertManager.sendAlert(managerId, "The bid for product " + currentBid.getProductId() + " in store " + this.storeId + " has been rejected already");
         currentBid.approve(managerId);
         Set<Integer> managers = storePermission.getAllUsersWithPermission("purchaseProposalSubmit");
 
         if (currentBid.approvedByAll(managers)) {
-            addToUserCart.apply(currentBid.getUserId(), storeId, currentBid.getProductId());
+            BroadCaster.broadcast(currentBid.getUserId(),
+                    "Your bid for product " + productRepository.getProductById(currentBid.getProductId()).getName()
+                            + " in store " + this.storeName + " approved!, the item added to your cart");
+            addToUserCart.apply(currentBid.getUserId(), storeId, currentBid.getProductId(), currentBid.getAmount(), currentBid.getNewProductPrice());
             //todo send alert to the user that his bid has been approved
         }
     }
 
     @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
+    @AuctionPermission
     @DefaultOwnerFunctionality
     public void purchaseProposalReject(int managerId, int bidId) throws NoPermissionException {
         /*
@@ -402,6 +445,11 @@ public class Store {
         if (!this.storePermission.checkPermission(managerId, hidden))//the user should be loggedIn with permissions
             throw new NoPermissionException("User " + managerId + " has no permission to reject a purchase proposal in the store: " + this.storeId);
         BID currentBid = bidRepository.getBID(bidId).orElseThrow(() -> new IllegalArgumentException("There is no such bid for store " + this.storeId));
+        if (currentBid.isRejected())
+            BroadCaster.broadcast(managerId,
+                "The bid for product " + productRepository.getProductById(currentBid.getProductId()).getName()
+                        + " has been rejected already");
+
         currentBid.reject();//good for concurrency edge cases
         bidRepository.removeBID(bidId);
         //todo send alert to the user that his bid has been rejected
@@ -409,7 +457,7 @@ public class Store {
 
     //only members of the store can create an auction purchase
     @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
+    @AuctionPermission
     @DefaultOwnerFunctionality
     public void auctionPurchase(int userId, int productId, double newPrice) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
@@ -418,7 +466,7 @@ public class Store {
     }
 
     @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
+    @AuctionPermission
     @DefaultOwnerFunctionality
     public PublicAuctionInfo getAuctionInfo(int userId, int productId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
@@ -428,7 +476,7 @@ public class Store {
     }
 
     @DefaultFounderFunctionality
-    @DefaultManagerFunctionality
+    @AuctionPermission
     @DefaultOwnerFunctionality
     public void createAuctionForProduct(int storeManagerId, int productId, double startingPrice, LocalDateTime lastDate) throws NoPermissionException {
         if (!this.storePermission.checkPermission(storeManagerId, hidden))
@@ -438,7 +486,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
-    @DefaultManagerFunctionality
+    @AuctionPermission
     public void endAuctionForProduct(int storeManagerId, int productId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(storeManagerId, hidden))
             throw new NoPermissionException("User " + storeManagerId + " has no permission to remove an auction in the store: " + this.storeId);
@@ -459,6 +507,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public void setProductName(int userId, int productId, String name) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to set product name in the store: " + this.storeId);
@@ -471,6 +520,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public void setProductCategory(int userId, int productId, String category) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to set product category in the store: " + this.storeId);
@@ -483,6 +533,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public void setProductPrice(int userId, int productId, double price) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to set product price in the store: " + this.storeId);
@@ -495,6 +546,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public void setProductStockQuantity(int userId, int productId, int quantity) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to set product stock quantity in the store: " + this.storeId);
@@ -507,6 +559,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public void setProductDescription(int userId, int productId, String description) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to set product description in the store: " + this.storeId);
@@ -519,6 +572,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @StockPermission
     public void removeProduct(int userId, int productId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to remove product in the store: " + this.storeId);
@@ -555,6 +609,7 @@ public class Store {
     }
 
     @DefaultFounderFunctionality
+    @FounderOnlyPermission
     public synchronized void hideStore(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to hide the store: " + this.storeId);
@@ -582,6 +637,7 @@ public class Store {
     }
 
     @DefaultFounderFunctionality
+    @FounderOnlyPermission
     public synchronized void unhideStore(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to unhide the store: " + this.storeId);
@@ -613,6 +669,8 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @HistoryPermission
+    @DefaultManagerFunctionality
     public List<PurchaseHistory> getStorePurchaseHistory(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to get the store purchase history: " + this.storeId);
@@ -649,6 +707,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public void setPurchasePolicyCondition(int userId, int conditionId) throws NoPermissionException{
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to set the purchase policy in the store : " + this.storeId);
@@ -658,6 +717,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addORCondition(int userId, int condition1, int condition2) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add OR condition in the store: " + this.storeId);
@@ -667,6 +727,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addANDCondition(int userId, int condition1, int condition2) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add AND condition in the store: " + this.storeId);
@@ -676,6 +737,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addXORCondition(int userId, int condition1, int condition2) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add XOR condition in the store: " + this.storeId);
@@ -685,6 +747,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addIMPLYCondition(int userId, int condition1, int condition2) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add IMPLY condition in the store: " + this.storeId);
@@ -694,6 +757,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStorePriceCondition(int userId, double lowerBound, double upperBound) throws NoPermissionException{
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store price condition in the store: " + this.storeId);
@@ -703,6 +767,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStorePriceCondition(int userId, double lowerBound) throws NoPermissionException{
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store price condition in the store: " + this.storeId);
@@ -712,6 +777,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStoreQuantityCondition(int userId, int lowerBound, int upperBound) throws NoPermissionException{
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store quantity condition in the store: " + this.storeId);
@@ -721,6 +787,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStoreQuantityCondition(int userId, int lowerBound) throws NoPermissionException{
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store quantity condition in the store: " + this.storeId);
@@ -729,6 +796,7 @@ public class Store {
     }
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryPriceCondition(int userId, String category, double lowerBound, double upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category price condition in the store: " + this.storeId);
@@ -738,6 +806,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryPriceCondition(int userId, String category, double lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category price condition in the store: " + this.storeId);
@@ -747,6 +816,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryQuantityCondition(int userId, String category, int lowerBound, int upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category quantity condition in the store: " + this.storeId);
@@ -756,6 +826,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryQuantityCondition(int userId, String category, int lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category quantity condition in the store: " + this.storeId);
@@ -765,6 +836,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addDateCondition(int userId, LocalDateTime lowerBound, LocalDateTime upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add date condition in the store: " + this.storeId);
@@ -774,6 +846,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addDateCondition(int userId, LocalDateTime lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add date condition in the store: " + this.storeId);
@@ -783,6 +856,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductPriceCondition(int userId, int productId, double lowerBound, double upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product price condition in the store: " + this.storeId);
@@ -792,6 +866,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductPriceCondition(int userId, int productId, double lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product price condition in the store: " + this.storeId);
@@ -801,6 +876,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductQuantityCondition(int userId, int productId, int lowerBound, int upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product quantity condition in the store: " + this.storeId);
@@ -810,6 +886,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductQuantityCondition(int userId, int productId, int lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product quantity condition in the store: " + this.storeId);
@@ -819,6 +896,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addTimeCondition(int userId, LocalDateTime lowerBound, LocalDateTime upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to time condition in the store: " + this.storeId);
@@ -828,6 +906,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addTimeCondition(int userId, LocalDateTime lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " time condition in the store: " + this.storeId);
@@ -837,6 +916,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addUserAgeCondition(int userId, int lowerBound, int upperBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " user age condition in the store: " + this.storeId);
@@ -846,6 +926,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addUserAgeCondition(int userId, int lowerBound) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " user age condition in the store: " + this.storeId);
@@ -858,6 +939,7 @@ public class Store {
      */
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStoreDiscount(int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String coupon) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store discount in the store: " + this.storeId);
@@ -867,6 +949,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStoreDiscount(int userId, double discountPercentage, LocalDate expirationDate, String coupon) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store discount in the store: " + this.storeId);
@@ -876,6 +959,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStoreDiscount(int userId, int conditionId, double discountPercentage, LocalDate expirationDate) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store discount in the store: " + this.storeId);
@@ -885,6 +969,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addStoreDiscount(int userId, double discountPercentage, LocalDate expirationDate) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add store discount in the store: " + this.storeId);
@@ -894,6 +979,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryDiscount(int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String category, String coupon) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category discount in the store: " + this.storeId);
@@ -903,6 +989,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryDiscount(int userId, double discountPercentage, LocalDate expirationDate, String category, String coupon) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category discount in the store: " + this.storeId);
@@ -912,6 +999,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryDiscount(int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String category) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category discount in the store: " + this.storeId);
@@ -922,6 +1010,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addCategoryDiscount(int userId, double discountPercentage, LocalDate expirationDate, String category) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add category discount in the store: " + this.storeId);
@@ -931,6 +1020,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductDiscount(int userId, int conditionId, double discountPercentage, LocalDate expirationDate, int productId, String coupon) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product discount in the store: " + this.storeId);
@@ -940,6 +1030,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductDiscount(int userId, double discountPercentage, LocalDate expirationDate, int productId, String coupon) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product discount in the store: " + this.storeId);
@@ -949,6 +1040,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductDiscount(int userId, int conditionId, double discountPercentage, LocalDate expirationDate, int productId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product discount in the store: " + this.storeId);
@@ -958,6 +1050,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public int addProductDiscount(int userId, double discountPercentage, LocalDate expirationDate, int productId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add product discount in the store: " + this.storeId);
@@ -967,6 +1060,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public List<StoreDiscount> getStoreDiscounts(int userId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to get store discounts in the store: " + this.storeId);
@@ -976,6 +1070,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public StoreDiscount getDiscount(int userId, int discountId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to get store discounts in the store: " + this.storeId);
@@ -985,6 +1080,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public void removeDiscount(int userId, int discountId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to remove store discounts in the store: " + this.storeId);
@@ -997,6 +1093,7 @@ public class Store {
      */
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public void addDiscountAsRoot(int userId, int discountId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add discount as root in the store: " + this.storeId);
@@ -1005,6 +1102,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public void addDiscountToXORRoot(int userId, int discountId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add discount to XOR root in the store: " + this.storeId);
@@ -1013,6 +1111,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public void addDiscountToMAXRoot(int userId, int discountId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add discount to MAX root in the store: " + this.storeId);
@@ -1022,6 +1121,7 @@ public class Store {
 
     @DefaultFounderFunctionality
     @DefaultOwnerFunctionality
+    @PoliciesPermission
     public void addDiscountToADDRoot(int userId, int discountId) throws NoPermissionException {
         if (!this.storePermission.checkPermission(userId, hidden))
             throw new NoPermissionException("User " + userId + " has no permission to add discount to ADD root in the store: " + this.storeId);
@@ -1029,4 +1129,39 @@ public class Store {
 
     }
 
+    @DefaultFounderFunctionality
+    @DefaultOwnerFunctionality
+    @PoliciesPermission
+    public DiscountAccumulationTreeInfo getDiscountAccumulationTree(int userId) throws NoPermissionException {
+        if(!this.storePermission.checkPermission(userId, hidden))
+            throw new NoPermissionException("User " + userId + " has no permission to get discount accumulation tree in the store: " + this.storeId);
+        return discountPolicy.getDiscountAccumulationTree();
+    }
+
+    @DefaultFounderFunctionality
+    @DefaultOwnerFunctionality
+    @PoliciesPermission
+    public void deleteStoreAccumulationTree(int userId) throws NoPermissionException {
+        if(!this.storePermission.checkPermission(userId, hidden))
+            throw new NoPermissionException("User " + userId + " has no permission to delete discount accumulation tree in the store: " + this.storeId);
+        discountPolicy.deleteStoreAccumulationTree();
+    }
+
+    @DefaultFounderFunctionality
+    @DefaultOwnerFunctionality
+    @PoliciesPermission
+    public Condition getPurchasePolicy(int userId) throws NoPermissionException {
+        if(!this.storePermission.checkPermission(userId, hidden))
+            throw new NoPermissionException("User " + userId + " has no permission to get the purchase policy in the store: " + this.storeId);
+        return purchasePolicy.getRootCondition();
+    }
+
+    @DefaultFounderFunctionality
+    @DefaultOwnerFunctionality
+    @PoliciesPermission
+    public void resetPurchasePolicy(int userId) throws NoPermissionException {
+        if(!this.storePermission.checkPermission(userId, hidden))
+            throw new NoPermissionException("User " + userId + " has no permission to reset the purchase policy in the store: " + this.storeId);
+        purchasePolicy.resetPurchasePolicy();
+    }
 }
