@@ -6,6 +6,9 @@ import BGU.Group13B.backend.User.UserInfo;
 import BGU.Group13B.backend.storePackage.newDiscoutns.Bounder;
 import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.Condition;
 import BGU.Group13B.backend.storePackage.purchaseBounders.PurchaseExceedsPolicyException;
+import BGU.Group13B.frontEnd.components.policyComponent.conditionEntities.ConditionEntity;
+import BGU.Group13B.frontEnd.components.policyComponent.conditionEntities.LogicalConditions.LogicalConditionEntity;
+import BGU.Group13B.frontEnd.components.policyComponent.conditionEntities.leaves.CategoryPriceConditionEntity;
 
 public class CategoryPriceCondition extends Condition {
 
@@ -13,17 +16,17 @@ public class CategoryPriceCondition extends Condition {
     private final Bounder<Double> priceBounder;
 
 
-    public CategoryPriceCondition(int conditionId, String category, double lowerBound, double upperBound){
+    public CategoryPriceCondition(int conditionId, String category, double lowerBound, double upperBound) {
         super(conditionId);
         this.category = category;
         this.priceBounder = new Bounder<>(lowerBound, upperBound);
     }
 
     /**
-     * @param category      product id
-     * @param lowerBound    dirbalek call this with upperbound
+     * @param category   product id
+     * @param lowerBound dirbalek call this with upperbound
      */
-    public CategoryPriceCondition(int conditionId, String category, double lowerBound){
+    public CategoryPriceCondition(int conditionId, String category, double lowerBound) {
         super(conditionId);
         this.category = category;
         this.priceBounder = new Bounder<>(lowerBound);
@@ -34,13 +37,22 @@ public class CategoryPriceCondition extends Condition {
         //sum products quantity in category
         double price = basketInfo.basketProducts().stream().
                 filter(p -> p.getCategory().equals(category)).
-                map(BasketProduct::getPrice).reduce(0.0, Double::sum);
+                map(b -> b.getPrice() * b.getQuantity()).reduce(0.0, Double::sum);
 
-        if(!priceBounder.inBounds(price)){
+        if (!priceBounder.inBounds(price)) {
             throw new PurchaseExceedsPolicyException("category price must be " + priceBounder);
         }
     }
-    public String toString(){
+
+    @Override
+    public ConditionEntity convertToUiEntity(LogicalConditionEntity parent) {
+        if(priceBounder.hasUpperBound()){
+            return new CategoryPriceConditionEntity(category, parent, priceBounder.getLowerBound(), priceBounder.getUpperBound());
+        }
+        return new CategoryPriceConditionEntity(category, parent, priceBounder.getLowerBound());
+    }
+
+    public String toString() {
         return "category: " + category + ", price: " + priceBounder.toString();
     }
 }
