@@ -3,7 +3,10 @@ package BGU.Group13B.frontEnd.components.views;
 import BGU.Group13B.backend.User.Message;
 import BGU.Group13B.backend.User.UserPermissions;
 import BGU.Group13B.backend.storePackage.WorkerCard;
+import BGU.Group13B.frontEnd.ResponseHandler;
 import BGU.Group13B.frontEnd.components.SessionToIdMapper;
+import BGU.Group13B.frontEnd.components.store.products.StoreProductsLayout;
+import BGU.Group13B.frontEnd.components.store.workers.StoreWorkersLayout;
 import BGU.Group13B.service.Response;
 import BGU.Group13B.service.Session;
 import BGU.Group13B.service.VoidResponse;
@@ -11,52 +14,31 @@ import BGU.Group13B.service.info.ProductInfo;
 import BGU.Group13B.service.info.StoreInfo;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.UI;
-import com.vaadin.flow.component.accordion.Accordion;
-import com.vaadin.flow.component.avatar.Avatar;
-import com.vaadin.flow.component.avatar.AvatarGroup;
-import com.vaadin.flow.component.avatar.AvatarGroup.AvatarGroupItem;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
-import com.vaadin.flow.component.combobox.ComboBox;
-import com.vaadin.flow.component.contextmenu.MenuItem;
-import com.vaadin.flow.component.contextmenu.SubMenu;
 import com.vaadin.flow.component.dialog.Dialog;
-import com.vaadin.flow.component.formlayout.FormLayout;
-import com.vaadin.flow.component.grid.Grid;
-import com.vaadin.flow.component.grid.GridVariant;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.H1;
-import com.vaadin.flow.component.html.H2;
 import com.vaadin.flow.component.html.Label;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.menubar.MenuBar;
-import com.vaadin.flow.component.menubar.MenuBarVariant;
 import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.notification.NotificationVariant;
-import com.vaadin.flow.component.orderedlayout.FlexComponent;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.progressbar.ProgressBar;
 import com.vaadin.flow.component.textfield.*;
-import com.vaadin.flow.router.BeforeEvent;
-import com.vaadin.flow.router.HasUrlParameter;
-import com.vaadin.flow.router.PageTitle;
-import com.vaadin.flow.router.Route;
+import com.vaadin.flow.router.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.vaadin.lineawesome.LineAwesomeIcon;
 
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Locale;
+import java.util.*;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 
 @PageTitle("Store")
 @Route(value = "store", layout = MainLayout.class)
-public class StoreView extends VerticalLayout implements HasUrlParameter<Integer> {
+public class StoreView extends VerticalLayout implements HasUrlParameter<Integer>, BeforeEnterObserver, ResponseHandler {
 
     private final Session session;
     private final int userId = SessionToIdMapper.getInstance().getCurrentSessionId();
@@ -64,15 +46,11 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
     private int storeId = -1;
     private StoreInfo storeInfo;
     private List<WorkerCard> workers;
-    private HashMap<UserPermissions.StoreRole, List<WorkerCard>> rolesToWorkers;
     private HashMap<Integer, String> userIdToUsername;
-    private List<ProductInfo> products;
-    private HashMap<String, List<ProductInfo>> categoriesToProducts;
-    private String[] rolesToAdd = {"Owner", "Manager"};
-    private List<String> allPermissions;
-    private List<String> founderPermissions;
-    private List<String> ownerPermissions;
-    private List<String> managerPermissions;
+    private Set<ProductInfo> products;
+    private boolean isAdmin = false;
+    private boolean isHidden = false;
+    private boolean terminate = false;
 
     //components
     private VerticalLayout headerLayout;
@@ -81,21 +59,19 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
     private VerticalLayout scoreLayout;
     private ProgressBar scoreBar;
     private Div scoreLabel;
+    private Icon hiddenStoreIcon;
     private HorizontalLayout bodyLayout;
-    private Accordion categoriesAccordion;
-    private HashMap<String, Grid<ProductInfo>> categoriesToGrids;
-    private Button addProductButton;
-    private VerticalLayout workersLayout;
-    private HorizontalLayout workersHeaderLayout;
-    private VerticalLayout productsLayout;
-    private HorizontalLayout productsHeaderLayout;
-    private Accordion rolesAccordion;
-    private HashMap<String, MenuBar> rolesToAvatarGroupMenuBars;
-    private Button addWorkerButton;
     private Button hideStoreButton;
+    private Button unhideStoreButton;
     private Button deleteStoreButton;
+    private Button storePurchaseHistoryButton;
+    private Button manageDiscountsButton;
+    private Button managePurchasePolicyButton;
     private HorizontalLayout bottomButtonsLayout;
     private Button storeMessagesButton;
+    private StoreProductsLayout storeProductsLayout;
+    private StoreWorkersLayout storeWorkersLayout;
+
 
 
 
@@ -108,15 +84,47 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
     private void start(){
         this.removeAll();
         init_dataFields();
-        demoData();
+        //demoData();
+        getData();
+        if(terminate){
+            return;
+        }
         getCurrentWorkerCard();
         init_components();
 
         init_header();
         init_body();
         init_bottomButtons();
+        setInvisibleBasedOnPermissions();
         setSizeFull();
     }
+
+    private void setInvisibleBasedOnPermissions(){
+        if(!currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.HISTORY)){
+            storePurchaseHistoryButton.setVisible(false);
+        }
+
+        if(!currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.POLICIES)){
+            managePurchasePolicyButton.setVisible(false);
+            manageDiscountsButton.setVisible(false);
+        }
+
+        if(!currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.FONLY)){
+            hideStoreButton.setVisible(false);
+            unhideStoreButton.setVisible(false);
+        }
+
+        if(!currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.WORKERS_INFO)){
+            storeWorkersLayout.setVisible(false);
+        }
+
+        if(!isAdmin){
+            deleteStoreButton.setVisible(false);
+        }
+
+
+    }
+
 
     private void init_components(){
         headerLayout = new VerticalLayout();
@@ -126,21 +134,15 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
         scoreBar = new ProgressBar();
         scoreLabel = new Div();
         bodyLayout = new HorizontalLayout();
-        categoriesAccordion = new Accordion();
-        categoriesToGrids = new HashMap<>();
-        workersLayout = new VerticalLayout();
-        workersHeaderLayout = new HorizontalLayout();
-        productsLayout = new VerticalLayout();
-        productsHeaderLayout = new HorizontalLayout();
-        rolesAccordion = new Accordion();
-        rolesToAvatarGroupMenuBars = new HashMap<>();
         bottomButtonsLayout = new HorizontalLayout();
-
-        addProductButton = new Button(new Icon(VaadinIcon.PLUS));
-        addWorkerButton = new Button(new Icon(VaadinIcon.PLUS));
         hideStoreButton = new Button("Hide Store");
+        unhideStoreButton = new Button("Reopen Store");
         deleteStoreButton = new Button("Delete Store");
+        storePurchaseHistoryButton = new Button("Store Purchase History");
+        manageDiscountsButton = new Button("Manage Discounts");
+        managePurchasePolicyButton = new Button("Manage Purchase Policy");
         storeMessagesButton = messageStoreDialog();
+        hiddenStoreIcon = new Icon(VaadinIcon.EYE_SLASH);
     }
 
     private void init_header(){
@@ -174,12 +176,14 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
         // Define header name and score layout
         H1 storeName = new H1(storeInfo.storeName());
         storeName.getStyle().set("margin-bottom", "0");
+        hiddenStoreIcon.setVisible(isHidden);
+        hideStoreButton.setVisible(!isHidden);
+        unhideStoreButton.setVisible(isHidden);
         storeMessagesButton.getStyle().set("margin-left", "auto");
-        header_name_score.add(storeName, scoreLayout, storeMessagesButton);
+        header_name_score.add(hiddenStoreIcon, storeName, scoreLayout, storeMessagesButton);
         header_name_score.setSpacing(false);
         header_name_score.getStyle().set("margin-bottom", "0");
         header_name_score.setWidthFull();
-
 
         // Define category label
         Label categoryLabel = new Label("Category: " + storeInfo.category());
@@ -195,234 +199,159 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
 
 
     private void init_body(){
-        init_products_section();
-        init_workers_section();
-        bodyLayout.add(productsLayout, workersLayout);
+        storeProductsLayout = new StoreProductsLayout(userId, storeId, session, products, currentWorker);
+        storeWorkersLayout = new StoreWorkersLayout(userId, storeId, session, workers, userIdToUsername, currentWorker);
+        bodyLayout.add(storeProductsLayout, storeWorkersLayout);
         bodyLayout.setSizeFull();
         add(bodyLayout);
     }
 
     private void init_bottomButtons() {
-        hideStoreButton.addClickListener(e -> {
-            handleResponse(session.hideStore(userId, storeId));
+
+        manageDiscountsButton.addClickListener(e -> {
+            if(currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.POLICIES)){
+                navigate("manageDiscounts/" + storeId);
+            }
+            else{
+                notifyWarning("You don't have permission to manage store discounts");
+            }
         });
+
+        managePurchasePolicyButton.addClickListener(e -> {
+            if(currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.POLICIES)){
+                navigate("managePurchasePolicy/" + storeId);
+            }
+            else{
+                notifyWarning("You don't have permission to manage store purchase policy");
+            }
+        });
+
+        storePurchaseHistoryButton.addClickListener(e-> {
+            if(currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.HISTORY)){
+                navigate("purchaseHistory/" + storeId); //TODO: change, ask eden
+            }
+            else{
+                notifyWarning("You don't have permission to view store purchase history");
+            }
+        });
+
+        hideStoreButton.addClickListener(e -> {
+            if(isHidden){
+                notifyWarning("Store is already hidden");
+                return;
+            }
+            if(handleResponse(session.hideStore(userId, storeId)) != null){
+                isHidden = true;
+                hiddenStoreIcon.setVisible(true);
+                hideStoreButton.setVisible(false);
+                unhideStoreButton.setVisible(true);
+                notifySuccess("Store has been hidden successfully");
+            }
+
+        });
+
+        unhideStoreButton.addClickListener(e -> {
+            if(!isHidden){
+                notifyWarning("Store is not hidden");
+                return;
+            }
+            if(handleResponse(session.unhideStore(userId, storeId)) != null){
+                isHidden = false;
+                hiddenStoreIcon.setVisible(false);
+                hideStoreButton.setVisible(true);
+                unhideStoreButton.setVisible(false);
+                notifySuccess("Store has been reopened successfully");
+            }
+        });
+
         deleteStoreButton.addClickListener(e -> {
-            Notification.show("not implemented yet in GUI :(");
+            if(handleResponse(session.deleteStore(userId, storeId)) != null){
+                notifySuccess("Store deleted successfully");
+                UI.getCurrent().navigate(HomeView.class);
+            }
+
         });
         deleteStoreButton.addThemeVariants(ButtonVariant.LUMO_ERROR);
 
-        bottomButtonsLayout.add(hideStoreButton, deleteStoreButton);
+        bottomButtonsLayout.add(manageDiscountsButton, managePurchasePolicyButton, storePurchaseHistoryButton, hideStoreButton, unhideStoreButton, deleteStoreButton);
         bottomButtonsLayout.getStyle().set("margin-top", "auto");
         add(bottomButtonsLayout);
     }
 
-    private void init_products_section(){
-        for(String category : categoriesToProducts.keySet()){
-            Grid<ProductInfo> grid = new Grid<>();
-            grid.setItems(categoriesToProducts.get(category));
-            grid.addColumn(ProductInfo::name).setHeader("Name");
-            grid.addColumn(ProductInfo::price).setHeader("Price");
-            grid.addColumn(ProductInfo::stockQuantity).setHeader("Stock");
-            grid.addColumn(p -> getRoundedScore(p.score())).setHeader("Rating");
-            categoriesToGrids.put(category, grid);
-
-            //styling
-            grid.getStyle().set("margin", "10px");
-            grid.setAllRowsVisible(true);
-            grid.setMaxWidth("960px");
-            grid.addThemeVariants(GridVariant.LUMO_ROW_STRIPES);
-            grid.addThemeVariants(GridVariant.LUMO_COLUMN_BORDERS);
-
-            //actions
-            grid.addItemDoubleClickListener(e -> {
-                ProductInfo productInfo = e.getItem();
-                navigate("product/" + productInfo.productId() + "/" + productInfo.storeId());
-            });
-        }
-
-        for(String category : categoriesToGrids.keySet()){
-            categoriesAccordion.add(category, categoriesToGrids.get(category));
-        }
-
-        addProductButton.addThemeVariants(ButtonVariant.LUMO_ICON);
-        addProductButton.addThemeVariants(ButtonVariant.LUMO_SMALL);
-        productsHeaderLayout.add(new H2("Products"), addProductButton);
-        productsHeaderLayout.setAlignItems(Alignment.CENTER);
-        productsLayout.add(productsHeaderLayout, categoriesAccordion);
-
-        //button action
-        addProductButton.addClickListener(e -> addProductDialog());
-    }
-
-    private void init_workers_section(){
-        for(UserPermissions.StoreRole role : rolesToWorkers.keySet()){
-            MenuBar menuBar = new MenuBar();
-            rolesToAvatarGroupMenuBars.put(roleToStringTitle(role), menuBar);
-            for(WorkerCard worker : rolesToWorkers.get(role)){
-                Avatar avatar = new Avatar(getUsername(worker));
-                menuBar.addThemeVariants(MenuBarVariant.LUMO_TERTIARY_INLINE);
-
-                MenuItem menuAvatarItem = menuBar.addItem(avatar);
-                SubMenu subMenu = menuAvatarItem.getSubMenu();
-                MenuItem profile = subMenu.addItem("Profile");
-                MenuItem removeWorker = subMenu.addItem("Remove Worker");
-
-                //actions
-                profile.addClickListener(e -> workerProfileDialog(worker));
-                removeWorker.addClickListener(e -> Notification.show("not implemented yet in GUI :("));
-            }
-        }
-
-
-        //ensure founder is first
-        if(rolesToAvatarGroupMenuBars.containsKey("Founder")){
-            rolesAccordion.add("Founder", rolesToAvatarGroupMenuBars.get("Founder"));
-        }
-
-        for(String role : rolesToAvatarGroupMenuBars.keySet()){
-            if(!role.equals("Founder")){
-                rolesAccordion.add(rolesToAvatarGroupMenuBars.get(role).getItems().size() > 1 ? role + "s" : role,
-                        rolesToAvatarGroupMenuBars.get(role));
-            }
-        }
-
-        addWorkerButton.addClickListener(e -> addWorkerDialog());
-
-        addWorkerButton.addThemeVariants(ButtonVariant.LUMO_ICON);
-        addWorkerButton.addThemeVariants(ButtonVariant.LUMO_SMALL);
-        workersHeaderLayout.add(new H2("Workers"), addWorkerButton);
-        workersHeaderLayout.setAlignItems(Alignment.CENTER);
-        workersLayout.add(workersHeaderLayout, rolesAccordion);
-
-        //uncomment to align workers to the right
-        workersLayout.getStyle().set("margin-left", "auto");
-        workersLayout.setMaxWidth("200px");
-
-
-    }
-
-    private void workerProfileDialog(WorkerCard workerCard){
-        Dialog dialog = new Dialog();
-        dialog.setWidth("350px");
-        dialog.setHeight("550px");
-        dialog.setCloseOnEsc(true);
-        dialog.setCloseOnOutsideClick(true);
-
-        H2 dialogTitle = new H2("Worker Profile");
-        Div workerName = new Div();
-        workerName.setText("Name: " + getUsername(workerCard));
-        Div workerRole = new Div();
-        workerRole.setText("Role: " + roleToStringTitle(workerCard.storeRole()));
-        Grid<String> workerPermissions = new Grid<>();
-        workerPermissions.setItems(workerCard.userPermissions());
-        workerPermissions.addColumn(p -> p).setHeader("Permissions");
-        workerPermissions.setAllRowsVisible(true);
-        workerPermissions.addThemeVariants(GridVariant.LUMO_ROW_STRIPES);
-        workerPermissions.addThemeVariants(GridVariant.LUMO_COLUMN_BORDERS);
-
-
-        VerticalLayout dialogLayout = new VerticalLayout();
-        dialogLayout.add(dialogTitle, workerName, workerRole, workerPermissions);
-        dialogLayout.setAlignItems(Alignment.STRETCH);
-
-        dialog.add(dialogLayout);
-        dialog.open();
-
-
-
-
-
-    }
-
-    private void addProductDialog(){
-        Dialog dialog = new Dialog();
-        dialog.setWidth("350px");
-        dialog.setHeight("550px");
-        dialog.setCloseOnEsc(true);
-        dialog.setCloseOnOutsideClick(true);
-
-        H2 dialogTitle = new H2("Add Product");
-        TextField productName = new TextField("Name");
-        TextField productCategory = new TextField("Category");
-        NumberField productPrice = new NumberField("Price");
-        IntegerField productStock = new IntegerField("Stock");
-        TextArea productDescription = new TextArea("Description");
-        productDescription.getStyle().set("max-height", "150px");
-        productDescription.getStyle().set("min-height", "150px");
-
-        productPrice.setValue(0.0);
-        productStock.setValue(0);
-        productPrice.setStepButtonsVisible(true);
-        productStock.setStepButtonsVisible(true);
-
-
-        Button confirmButton = new Button("Confirm");
-        confirmButton.addClickListener(e2 -> {
-            handleResponseAndRefresh(session.addProduct(userId, storeId, productName.getValue(), productCategory.getValue(),
-                    productPrice.getValue(), productStock.getValue(), productDescription.getValue()), "store/" + storeId);
-            dialog.close();
-        });
-
-
-        VerticalLayout dialogLayout = new VerticalLayout();
-        dialogLayout.add(dialogTitle, productName, productPrice, productStock, productDescription, confirmButton);
-        dialogLayout.setSpacing(false);
-        dialogLayout.setAlignItems(Alignment.STRETCH);
-
-        dialog.add(dialogLayout);
-        dialog.open();
-
-    }
-
-    private void addWorkerDialog(){
-        Dialog dialog = new Dialog();
-        dialog.setWidth("350px");
-        dialog.setHeight("300px");
-        dialog.setCloseOnEsc(true);
-        dialog.setCloseOnOutsideClick(true);
-
-        H2 dialogTitle = new H2("Add Worker");
-        TextField workerName = new TextField("Username");
-        ComboBox<String> workerRole = new ComboBox<>("Role");
-        workerRole.setItems(rolesToAdd);
-
-        Button confirmButton = new Button("Confirm");
-        confirmButton.addClickListener(e2 -> {
-            Notification.show("not implemented yet in GUI :(");
-            dialog.close();
-        });
-
-        VerticalLayout dialogLayout = new VerticalLayout();
-        dialogLayout.add(dialogTitle, workerName, workerRole, confirmButton);
-        dialogLayout.setSpacing(false);
-        dialogLayout.setAlignItems(Alignment.STRETCH);
-
-        dialog.add(dialogLayout);
-        dialog.open();
-    }
-
     private void getData(){
+        //isAdmin
+        isAdmin = handleResponse(session.isAdmin(userId), "");
+
+        //isHidden
+        isHidden = handleResponse(session.isStoreHidden(storeId), "");
+
+        //products
+        products = handleResponse(session.getAllStoreProductsInfo(userId, storeId), "");
+
+        //workers
+        workers = handleResponse(session.getStoreWorkersInfo(1 /*mafhhhiiddd*/, storeId), "");
+
+        //userIdToUsername
+        List<Integer> userIds = workers.stream().map(WorkerCard::userId).collect(Collectors.toList());
+        userIdToUsername = handleResponse(session.getUserIdsToUsernamesMapper(userIds), "");
+    }
+
+    private double getRoundedScore(double score){
+        return Math.round(score * 10.0) / 10.0;
+    }
+
+
+    private void init_dataFields(){
+        userIdToUsername = new HashMap<>();
 
     }
 
-    private void init_categoriesToProducts(){
-        categoriesToProducts = new HashMap<>();
-        for(ProductInfo product : products){
-            if(!categoriesToProducts.containsKey(product.category())){
-                categoriesToProducts.put(product.category(), new LinkedList<>());
-            }
-            categoriesToProducts.get(product.category()).add(product);
-        }
+    private void getCurrentWorkerCard(){
+        currentWorker = workers.stream().filter(worker -> worker.userId() == userId).findFirst().orElse(new WorkerCard(userId, null, Set.of()));
     }
 
-    private void init_rolesToWorkers(){
-        rolesToWorkers = new HashMap<>();
-        for(WorkerCard worker : workers){
-            if(!rolesToWorkers.containsKey(worker.storeRole())){
-                rolesToWorkers.put(worker.storeRole(), new LinkedList<>());
-            }
-            rolesToWorkers.get(worker.storeRole()).add(worker);
-        }
+
+    private void demoData(){
+        storeInfo = new StoreInfo(0, "SheepStore", "sheep", 3.9);
+
+        userIdToUsername.put(0, "lior");
+        userIdToUsername.put(1, "eden");
+        userIdToUsername.put(2, "eyal");
+        userIdToUsername.put(3, "shaun");
+        userIdToUsername.put(4, "tomer");
+        userIdToUsername.put(5, "yoav");
+
+        WorkerCard worker0 = new WorkerCard(0, UserPermissions.StoreRole.FOUNDER, Set.of(UserPermissions.IndividualPermission.values()));
+        WorkerCard worker1 = new WorkerCard(1, UserPermissions.StoreRole.OWNER, Set.of());
+        WorkerCard worker2 = new WorkerCard(2, UserPermissions.StoreRole.OWNER, Set.of());
+        WorkerCard worker3 = new WorkerCard(3, UserPermissions.StoreRole.MANAGER, Set.of());
+        WorkerCard worker4 = new WorkerCard(4, UserPermissions.StoreRole.MANAGER, Set.of());
+        WorkerCard worker5 = new WorkerCard(5, UserPermissions.StoreRole.MANAGER, Set.of());
+
+        workers = new LinkedList<>();
+        workers.add(worker0);
+        workers.add(worker1);
+        workers.add(worker2);
+        workers.add(worker3);
+        workers.add(worker4);
+        workers.add(worker5);
+
+        products = new LinkedHashSet<>();
+        products.add(new ProductInfo(0, 0, storeInfo.storeName(), "milk", "dairy", 5.0, 10, "milk description", 4.2F));
+        products.add(new ProductInfo(-1, 0, storeInfo.storeName(),"cheese", "dairy", 10.0, 20, "cheese description", 4.5F));
+        products.add(new ProductInfo(-2, 0, storeInfo.storeName(),"bread", "bakery", 3.0, 30, "bread description", 4.0F));
+        products.add(new ProductInfo(-3, 0, storeInfo.storeName(),"butter", "dairy", 7.0, 40, "butter description", 4.1F));
+        products.add(new ProductInfo(-4, 0, storeInfo.storeName(),"eggs", "dairy", 8.0, 50, "eggs description", 4.3F));
+        products.add(new ProductInfo(-5, 0, storeInfo.storeName(),"yogurt", "dairy", 6.0, 60, "yogurt description", 4.4F));
+        products.add(new ProductInfo(-6, 0, storeInfo.storeName(),"cake", "bakery", 12.0, 70, "cake description", 4.6F));
+        products.add(new ProductInfo(-7, 0, storeInfo.storeName(),"cookies", "bakery", 9.0, 80, "cookies description", 4.7F));
+        products.add(new ProductInfo(-8, 0, storeInfo.storeName(),"chocolate", "bakery", 11.0, 90, "chocolate description", 4.8F));
+        products.add(new ProductInfo(-9, 0, storeInfo.storeName(),"pizza", "bakery", 15.0, 100, "pizza description", 4.9F));
+        products.add(new ProductInfo(-10, 0, storeInfo.storeName(),"water", "drinks", 2.0, 110, "water description", 3.3F));
+        products.add(new ProductInfo(-11, 0, storeInfo.storeName(),"soda", "drinks", 3.0, 120, "soda description", 2.5F));
+        products.add(new ProductInfo(-12, 0, storeInfo.storeName(),"juice", "drinks", 4.0, 130, "juice description", 1.7F));
+        products.add(new ProductInfo(-13, 0, storeInfo.storeName(),"beer", "drinks", 5.0, 140, "beer description", 3.8F));
+        products.add(new ProductInfo(-14, 0, storeInfo.storeName(),"wine", "drinks", 6.0, 150, "wine description", 2.1F));
     }
 
     //TODO:lior - all the functions below are for store - please provide the store id and call to functions that checks if the store owner
@@ -469,7 +398,7 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
 
         });
 
-        if(currentWorker.userPermissions().contains("getUnreadMessages") /*if storeOwner - getUnreadMessages*/)
+        if(currentWorker.userPermissions().contains(UserPermissions.IndividualPermission.MESSAGES) /*if storeOwner - getUnreadMessages*/)
         {
             buttonsLayout.add(newMessagesStore, oldMessagesStore);
         }
@@ -496,7 +425,6 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
         VerticalLayout verticalDialogMessage = new VerticalLayout();
         TextArea messageBody = new TextArea();
         Button refreshMessagesButton = new Button("Refresh Messages");
-
         Response<Message> messageResponse = session.readReadMassageStore(userId,storeId);
         String message;
         if(messageResponse.getStatus()== Response.Status.SUCCESS) {
@@ -640,173 +568,24 @@ public class StoreView extends VerticalLayout implements HasUrlParameter<Integer
 
     }
 
-    private void init_userIdToUsername(){
-        //TODO: get from backend
-    }
-
-    private double getRoundedScore(double score){
-        return Math.round(score * 10.0) / 10.0;
-    }
-
-    private String getUsername(WorkerCard worker){
-        return userIdToUsername.get(worker.userId());
-    }
-
-    private String roleToStringTitle(UserPermissions.StoreRole role){
-        String roleString = role.toString();
-        String[] words = roleString.split("\\s+");
-        StringBuilder sb = new StringBuilder();
-        for (String word : words) {
-            sb.append(Character.toUpperCase(word.charAt(0)));
-            sb.append(word.substring(1).toLowerCase());
-            sb.append(" ");
+    @Override
+    public void setParameter(BeforeEvent beforeEvent, Integer storeIdParam) {
+        if(storeIdParam == null){
+            beforeEvent.rerouteTo("");
+            return;
         }
-        return sb.toString().trim();
-    }
 
-    private void init_dataFields(){
-        rolesToWorkers = new HashMap<>();
-        userIdToUsername = new HashMap<>();
-        categoriesToProducts = new HashMap<>();
-
-        //all permissions: [removeStoreDiscount, setStorePurchasePriceLowerBound, removeManager, removeProduct, setProductPurchasePriceLowerBound, addProduct, hideStore, setProductStockQuantity, setProductName, createAuctionForProduct, setStorePurchaseQuantityLowerBound, addManager, unhideStore, addStoreVisibleDiscount, getAuctionInfo, disallowPurchasePolicyConflicts, setProductPurchaseQuantityLowerBound, addStoreConditionalDiscount, setStorePurchasePriceBounds, setProductDescription, addProductVisibleDiscount, purchaseProposalSubmit, allowPurchasePolicyConflicts, setProductPurchasePriceUpperBound, purchaseProposalReject, removeProductDiscount, markAsCompleted, refreshMessages, auctionPurchase, getReadMessages, setProductPurchaseQuantityUpperBound, addOwner, setProductPurchaseQuantityBounds, getUnreadMessages, endAuctionForProduct, setProductPurchasePriceBounds, setStorePurchaseQuantityBounds, addStoreHiddenDiscount, addProductConditionalDiscount, removeOwner, setProductCategory, setStorePurchasePriceUpperBound, setStorePurchaseQuantityUpperBound, purchaseProposalApprove, addProductHiddenDiscount, getStoreWorkersInfo, setProductPrice, deleteStore]
-        allPermissions = List.of("removeStoreDiscount", "setStorePurchasePriceLowerBound", "removeManager", "removeProduct",
-                "setProductPurchasePriceLowerBound", "addProduct", "hideStore", "setProductStockQuantity",
-                "setProductName", "createAuctionForProduct", "setStorePurchaseQuantityLowerBound", "addManager",
-                "unhideStore", "addStoreVisibleDiscount", "getAuctionInfo", "disallowPurchasePolicyConflicts",
-                "setProductPurchaseQuantityLowerBound", "addStoreConditionalDiscount", "setStorePurchasePriceBounds",
-                "setProductDescription", "addProductVisibleDiscount", "purchaseProposalSubmit", "allowPurchasePolicyConflicts",
-                "setProductPurchasePriceUpperBound", "purchaseProposalReject", "removeProductDiscount", "markAsCompleted",
-                "refreshMessages", "auctionPurchase", "getReadMessages", "setProductPurchaseQuantityUpperBound", "addOwner",
-                "setProductPurchaseQuantityBounds", "getUnreadMessages", "endAuctionForProduct", "setProductPurchasePriceBounds",
-                "setStorePurchaseQuantityBounds", "addStoreHiddenDiscount", "addProductConditionalDiscount", "removeOwner",
-                "setProductCategory", "setStorePurchasePriceUpperBound", "setStorePurchaseQuantityUpperBound",
-                "purchaseProposalApprove", "addProductHiddenDiscount", "getStoreWorkersInfo", "setProductPrice", "deleteStore");
-
-        //founder permissions: [removeStoreDiscount, setStorePurchasePriceLowerBound, removeManager, removeProduct, setProductPurchasePriceLowerBound, addProduct, hideStore, setProductStockQuantity, setProductName, createAuctionForProduct, setStorePurchaseQuantityLowerBound, addManager, unhideStore, addStoreVisibleDiscount, getAuctionInfo, disallowPurchasePolicyConflicts, setProductPurchaseQuantityLowerBound, addStoreConditionalDiscount, setStorePurchasePriceBounds, setProductDescription, addProductVisibleDiscount, purchaseProposalSubmit, allowPurchasePolicyConflicts, setProductPurchasePriceUpperBound, purchaseProposalReject, removeProductDiscount, markAsCompleted, refreshMessages, auctionPurchase, getReadMessages, setProductPurchaseQuantityUpperBound, addOwner, setProductPurchaseQuantityBounds, getUnreadMessages, endAuctionForProduct, setProductPurchasePriceBounds, setStorePurchaseQuantityBounds, addStoreHiddenDiscount, addProductConditionalDiscount, removeOwner, setProductCategory, setStorePurchasePriceUpperBound, setStorePurchaseQuantityUpperBound, purchaseProposalApprove, addProductHiddenDiscount, getStoreWorkersInfo, setProductPrice]
-        founderPermissions = List.of(
-                "removeStoreDiscount", "setStorePurchasePriceLowerBound", "removeManager", "removeProduct",
-                "setProductPurchasePriceLowerBound", "addProduct", "hideStore", "setProductStockQuantity",
-                "setProductName", "createAuctionForProduct", "setStorePurchaseQuantityLowerBound", "addManager",
-                "unhideStore", "addStoreVisibleDiscount", "getAuctionInfo", "disallowPurchasePolicyConflicts",
-                "setProductPurchaseQuantityLowerBound", "addStoreConditionalDiscount", "setStorePurchasePriceBounds",
-                "setProductDescription", "addProductVisibleDiscount", "purchaseProposalSubmit", "allowPurchasePolicyConflicts",
-                "setProductPurchasePriceUpperBound", "purchaseProposalReject", "removeProductDiscount", "markAsCompleted",
-                "refreshMessages", "auctionPurchase", "getReadMessages", "setProductPurchaseQuantityUpperBound", "addOwner",
-                "setProductPurchaseQuantityBounds", "getUnreadMessages", "endAuctionForProduct", "setProductPurchasePriceBounds",
-                "setStorePurchaseQuantityBounds", "addStoreHiddenDiscount", "addProductConditionalDiscount", "removeOwner",
-                "setProductCategory", "setStorePurchasePriceUpperBound", "setStorePurchaseQuantityUpperBound",
-                "purchaseProposalApprove", "addProductHiddenDiscount", "getStoreWorkersInfo", "setProductPrice"
-        );
-
-        //owner permissions: [removeManager, removeProduct, addOwner, addProduct, getUnreadMessages, setProductStockQuantity, setProductName, createAuctionForProduct, endAuctionForProduct, addManager, getAuctionInfo, removeOwner, setProductCategory, setProductDescription, purchaseProposalSubmit, purchaseProposalApprove, purchaseProposalReject, getStoreWorkersInfo, markAsCompleted, setProductPrice, refreshMessages, auctionPurchase, getReadMessages]
-        ownerPermissions = List.of(
-                "removeManager", "removeProduct", "addOwner", "addProduct", "getUnreadMessages", "setProductStockQuantity",
-                "setProductName", "createAuctionForProduct", "endAuctionForProduct", "addManager", "getAuctionInfo",
-                "removeOwner", "setProductCategory", "setProductDescription", "purchaseProposalSubmit", "purchaseProposalApprove",
-                "purchaseProposalReject", "getStoreWorkersInfo", "markAsCompleted", "setProductPrice", "refreshMessages",
-                "auctionPurchase", "getReadMessages"
-        );
-
-        //manager permissions: [addProduct, purchaseProposalSubmit, createAuctionForProduct, endAuctionForProduct, purchaseProposalApprove, purchaseProposalReject, getAuctionInfo, auctionPurchase]
-        managerPermissions = List.of(
-                "addProduct", "purchaseProposalSubmit", "createAuctionForProduct", "endAuctionForProduct",
-                "purchaseProposalApprove", "purchaseProposalReject", "getAuctionInfo", "auctionPurchase"
-        );
-    }
-
-    private void getCurrentWorkerCard(){
-        currentWorker = workers.stream().filter(worker -> worker.userId() == userId).findFirst().orElse(null);
-        if(currentWorker == null){
-            Notification erorrNotification = new Notification("Error: could not find worker with id " + userId, 3000);
-            erorrNotification.addThemeVariants(NotificationVariant.LUMO_ERROR);
-            erorrNotification.open();
-            navigate("");
-        }
-    }
-
-
-    private void demoData(){
-        storeInfo = new StoreInfo(0, "SheepStore", "sheep", 3.9);
-
-        userIdToUsername.put(0, "lior");
-        userIdToUsername.put(1, "eden");
-        userIdToUsername.put(2, "eyal");
-        userIdToUsername.put(3, "shaun");
-        userIdToUsername.put(4, "tomer");
-        userIdToUsername.put(5, "yoav");
-
-        WorkerCard worker0 = new WorkerCard(0, UserPermissions.StoreRole.FOUNDER, founderPermissions);
-        WorkerCard worker1 = new WorkerCard(1, UserPermissions.StoreRole.OWNER, ownerPermissions);
-        WorkerCard worker2 = new WorkerCard(2, UserPermissions.StoreRole.OWNER, ownerPermissions);
-        WorkerCard worker3 = new WorkerCard(3, UserPermissions.StoreRole.MANAGER, managerPermissions);
-        WorkerCard worker4 = new WorkerCard(4, UserPermissions.StoreRole.MANAGER, managerPermissions);
-        WorkerCard worker5 = new WorkerCard(5, UserPermissions.StoreRole.MANAGER, managerPermissions);
-
-        workers = new LinkedList<>();
-        workers.add(worker0);
-        workers.add(worker1);
-        workers.add(worker2);
-        workers.add(worker3);
-        workers.add(worker4);
-        workers.add(worker5);
-
-        products = new LinkedList<>();
-        products.add(new ProductInfo(0, 0, storeInfo.storeName(), "milk", "dairy", 5.0, 10, "milk description", 4.2F));
-        products.add(new ProductInfo(-1, 0, storeInfo.storeName(),"cheese", "dairy", 10.0, 20, "cheese description", 4.5F));
-        products.add(new ProductInfo(-2, 0, storeInfo.storeName(),"bread", "bakery", 3.0, 30, "bread description", 4.0F));
-        products.add(new ProductInfo(-3, 0, storeInfo.storeName(),"butter", "dairy", 7.0, 40, "butter description", 4.1F));
-        products.add(new ProductInfo(-4, 0, storeInfo.storeName(),"eggs", "dairy", 8.0, 50, "eggs description", 4.3F));
-        products.add(new ProductInfo(-5, 0, storeInfo.storeName(),"yogurt", "dairy", 6.0, 60, "yogurt description", 4.4F));
-        products.add(new ProductInfo(-6, 0, storeInfo.storeName(),"cake", "bakery", 12.0, 70, "cake description", 4.6F));
-        products.add(new ProductInfo(-7, 0, storeInfo.storeName(),"cookies", "bakery", 9.0, 80, "cookies description", 4.7F));
-        products.add(new ProductInfo(-8, 0, storeInfo.storeName(),"chocolate", "bakery", 11.0, 90, "chocolate description", 4.8F));
-        products.add(new ProductInfo(-9, 0, storeInfo.storeName(),"pizza", "bakery", 15.0, 100, "pizza description", 4.9F));
-        products.add(new ProductInfo(-10, 0, storeInfo.storeName(),"water", "drinks", 2.0, 110, "water description", 3.3F));
-        products.add(new ProductInfo(-11, 0, storeInfo.storeName(),"soda", "drinks", 3.0, 120, "soda description", 2.5F));
-        products.add(new ProductInfo(-12, 0, storeInfo.storeName(),"juice", "drinks", 4.0, 130, "juice description", 1.7F));
-        products.add(new ProductInfo(-13, 0, storeInfo.storeName(),"beer", "drinks", 5.0, 140, "beer description", 3.8F));
-        products.add(new ProductInfo(-14, 0, storeInfo.storeName(),"wine", "drinks", 6.0, 150, "wine description", 2.1F));
-
-        init_categoriesToProducts();
-        init_rolesToWorkers();
-    }
-
-
-    private <T> T handleResponse(Response<T> response, String navInCaseOfFailure) {
-        if (response.didntSucceed()) {
-            Notification errorNotification = Notification.show("Error: " + response.getMessage());
-            errorNotification.addThemeVariants(NotificationVariant.LUMO_ERROR);
-            navigate(navInCaseOfFailure);
-        }
-        return response.getData();
-    }
-
-    private <T> T handleResponse(Response<T> response) {
-        return handleResponse(response, "");
-    }
-
-    private <T> T handleResponseAndRefresh(Response<T> response, String navInCaseOfFailure) {
-        if (response.didntSucceed()) {
-            Notification errorNotification = Notification.show("Error: " + response.getMessage());
-            errorNotification.addThemeVariants(NotificationVariant.LUMO_ERROR);
-            navigate(navInCaseOfFailure);
-            start();
-        }
-        return response.getData();
-    }
-
-    private <T> T handleResponseAndRefresh(Response<T> response) {
-        return handleResponseAndRefresh(response, "");
-    }
-
-    private void navigate(String nav) {
-        UI.getCurrent().navigate(nav);
+        this.storeId = storeIdParam;
     }
 
     @Override
-    public void setParameter(BeforeEvent beforeEvent, Integer storeIdParam) {
-        this.storeId = storeIdParam;
+    public void beforeEnter(BeforeEnterEvent beforeEnterEvent) {
+        storeInfo = handleResponse(session.getStoreInfo(userId, storeId));
+        if(storeInfo == null){
+            terminate = true;
+            beforeEnterEvent.rerouteTo(HomeView.class);
+            return;
+        }
         start();
     }
 }
