@@ -5,15 +5,19 @@ import BGU.Group13B.backend.System.SystemInfo;
 import BGU.Group13B.backend.User.Message;
 import BGU.Group13B.backend.User.PurchaseFailedException;
 import BGU.Group13B.backend.User.PurchaseHistory;
+import BGU.Group13B.backend.User.UserPermissions;
 import BGU.Group13B.backend.storePackage.Review;
 import BGU.Group13B.backend.storePackage.PublicAuctionInfo;
 import BGU.Group13B.backend.storePackage.WorkerCard;
 import BGU.Group13B.backend.storePackage.newDiscoutns.DiscountInfo;
+import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.Condition;
 import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.StoreDiscount;
 import BGU.Group13B.backend.storePackage.permissions.NoPermissionException;
+import BGU.Group13B.frontEnd.components.policyComponent.conditionEntities.ConditionEntity;
 import BGU.Group13B.service.entity.ReviewService;
 import BGU.Group13B.service.entity.ServiceBasketProduct;
 import BGU.Group13B.service.entity.ServiceProduct;
+import BGU.Group13B.service.info.DiscountAccumulationTreeInfo;
 import BGU.Group13B.service.info.StoreInfo;
 import BGU.Group13B.service.info.ProductInfo;
 
@@ -36,7 +40,7 @@ public interface ISession {
      * @param price         the product price
      * @param stockQuantity the stock quantity
      * @param description   the product description
-     * @return              the product id
+     * @return the product id
      */
     Response<Integer> addProduct(int userId, int storeId, String productName, String category, double price, int stockQuantity, String description);
 
@@ -44,11 +48,13 @@ public interface ISession {
      * #19
      * require 2.3
      * good for development delete this function, does not check if the item exists in a store
+     *
      * @param userId    the user id
      * @param storeId   the store id
      * @param productId the product id
+     * @return
      */
-    void addToCart(int userId, int storeId, int productId);
+    Response<VoidResponse> addToCart(int userId, int storeId, int productId);
 
     /**
      * #22
@@ -70,11 +76,11 @@ public interface ISession {
                                String/*store coupons*/ storeCoupon);
 
     Response<VoidResponse> purchaseProductCart(int userId, String creditCardNumber,
-                                                      String creditCardMonth, String creditCardYear,
-                                                      String creditCardHolderFirstName,
-                                                      String creditCardCVV, String id,
-                                                      String address, String city, String country,
-                                                      String zip);
+                                               String creditCardMonth, String creditCardYear,
+                                               String creditCardHolderFirstName,
+                                               String creditCardCVV, String id,
+                                               String address, String city, String country,
+                                               String zip);
 
 
     /**
@@ -86,8 +92,12 @@ public interface ISession {
      * @param productId     the product id
      * @param proposedPrice the proposed price
      */
-    void purchaseProposalSubmit(int userId, int storeId, int productId, double proposedPrice, int amount);
+    Response<VoidResponse> purchaseProposalSubmit(int userId, int storeId, int productId, double proposedPrice, int amount);
 
+
+    void purchaseProposalApprove(int managerId, int storeId, int productId) throws NoPermissionException;
+
+    void purchaseProposalReject(int storeId, int managerId, int bidId) throws NoPermissionException;
 
     /**
      * #51
@@ -189,7 +199,7 @@ public interface ISession {
      * @param minPrice - the minimum price of the product
      * @param maxPrice - the maximum price of the product
      */
-    Response<List<ProductInfo>> filterByPriceRange(int minPrice, int maxPrice);
+    Response<List<ProductInfo>> filterByPriceRange(double minPrice, double maxPrice);
 
     /**
      * #18
@@ -198,7 +208,7 @@ public interface ISession {
      * @param minRating - the minimum rating of the product
      * @param maxRating - the maximum rating of the product
      */
-    Response<List<ProductInfo>> filterByProductRank(int minRating, int maxRating);
+    Response<List<ProductInfo>> filterByProductRank(double minRating, double maxRating);
 
     /**
      * #18
@@ -215,7 +225,7 @@ public interface ISession {
      * @param minRating - the minimum rating of the store
      * @param maxRating - the maximum rating of the store
      */
-    Response<List<ProductInfo>> filterByStoreRank(int minRating, int maxRating);
+    Response<List<ProductInfo>> filterByStoreRank(double minRating, double maxRating);
 
     /**
      * #16
@@ -258,6 +268,7 @@ public interface ISession {
      * #19
      * require 2.3
      * good for production
+     *
      * @param userId    the user id
      * @param storeId   the store id
      * @param productId the product id
@@ -567,10 +578,10 @@ public interface ISession {
      * #32
      * require 4.1
      *
-     * @param userId        the user id
-     * @param storeId       the store id
-     * @param productId     the product id
-     * @param description   the product description
+     * @param userId      the user id
+     * @param storeId     the store id
+     * @param productId   the product id
+     * @param description the product description
      */
     Response<VoidResponse> setProductDescription(int userId, int storeId, int productId, String description);
 
@@ -725,10 +736,10 @@ public interface ISession {
     void exitSystemAsGuest(int userId);
 
     /**
-     * @param userId the user id
+     * @param userId  the user id
      * @param storeId the store id
      * @return the list of the products that the user failed buy from the store
-     * */
+     */
     List<Integer> getFailedProducts(int userId, int storeId);
 
 
@@ -740,7 +751,6 @@ public interface ISession {
 
 
     /**
-     *
      * @param userId the user id
      * @return all the user's associated stores (all stores that the user has a role in)
      */
@@ -751,8 +761,8 @@ public interface ISession {
      * #39
      * require 4.9
      *
-     * @param userId    the user id (only the founder can do this action)
-     * @param storeId   the store id that will be hidden
+     * @param userId  the user id (only the founder can do this action)
+     * @param storeId the store id that will be hidden
      */
     Response<VoidResponse> hideStore(int userId, int storeId);
 
@@ -760,8 +770,8 @@ public interface ISession {
      * #40
      * require 4.10
      *
-     * @param userId    the user id (only the founder can do this action)
-     * @param storeId   the store id that will be unhidden
+     * @param userId  the user id (only the founder can do this action)
+     * @param storeId the store id that will be unhidden
      */
     Response<VoidResponse> unhideStore(int userId, int storeId);
 
@@ -769,8 +779,8 @@ public interface ISession {
      * #45
      * require 6.1
      *
-     * @param userId    the user id (only an admin can do this action)
-     * @param storeId   the store id that will be deleted
+     * @param userId  the user id (only an admin can do this action)
+     * @param storeId the store id that will be deleted
      */
     Response<VoidResponse> deleteStore(int userId, int storeId);
 
@@ -781,23 +791,26 @@ public interface ISession {
      * @return the user's purchase history
      * @throws Exception
      */
-    Response<String> getUserPurchaseHistory(int userId);
-    Response<String> getUserPurchaseHistoryAsAdmin(int userId, int adminId);
+    Response<List<PurchaseHistory>> getUserPurchaseHistory(int userId);
+
+    Response<List<PurchaseHistory>>  getUserPurchaseHistoryAsAdmin(int userId, int adminId);
+
     Response<List<PurchaseHistory>> getStorePurchaseHistoryAsAdmin(int storeId, int adminId);
 
 
     Response<VoidResponse> fetchMessages(int userId);
 
     int getStoreFounder(int storeId);
+
     Pair<Double, List<ServiceBasketProduct>> startPurchaseBasketTransaction(int userId, List<String> coupons) throws PurchaseFailedException;
 
     /**
      * #43
      * require 4.13
      *
-     * @param userId    the user id
-     * @param storeId   the store id
-     * @return          the store purchase history
+     * @param userId  the user id
+     * @param storeId the store id
+     * @return the store purchase history
      */
     Response<List<PurchaseHistory>> getStorePurchaseHistory(int userId, int storeId);
 
@@ -805,10 +818,10 @@ public interface ISession {
      * #XX
      * require XXX?
      *
-     * @param userId the user id
+     * @param userId     the user id
      * @param newOwnerId the id of the new owner
-     * @param storeId the id of the store the new owner will own
-     *               as an owner/founder you can add other users as store owners in your store
+     * @param storeId    the id of the store the new owner will own
+     *                   as an owner/founder you can add other users as store owners in your store
      */
     Response<VoidResponse> addOwner(int userId, int newOwnerId, int storeId);
 
@@ -816,10 +829,10 @@ public interface ISession {
      * #XX
      * require XXX?
      *
-     * @param userId the user id
+     * @param userId        the user id
      * @param removeOwnerId the id of the owner to be removed
-     * @param storeId the id of the store the owner will be removed from
-     *               as an owner/founder you can remove owners you assigned yourself
+     * @param storeId       the id of the store the owner will be removed from
+     *                      as an owner/founder you can remove owners you assigned yourself
      */
     Response<VoidResponse> removeOwner(int userId, int removeOwnerId, int storeId);
 
@@ -827,10 +840,10 @@ public interface ISession {
      * #XX
      * require XXX?
      *
-     * @param userId the user id
+     * @param userId       the user id
      * @param newManagerId the id of the new manager
-     * @param storeId the id of the store the new manager will manage
-     *               as an owner/founder you can add other users as store managers in your store
+     * @param storeId      the id of the store the new manager will manage
+     *                     as an owner/founder you can add other users as store managers in your store
      */
     Response<VoidResponse> addManager(int userId, int newManagerId, int storeId);
 
@@ -838,10 +851,10 @@ public interface ISession {
      * #XX
      * require XXX?
      *
-     * @param userId the user id
+     * @param userId          the user id
      * @param removeManagerId the id of the manager to be removed
-     * @param storeId the id of the store the manager will be removed from
-     *               as an owner/founder you can remove managers you assigned yourself
+     * @param storeId         the id of the store the manager will be removed from
+     *                        as an owner/founder you can remove managers you assigned yourself
      */
     Response<VoidResponse> removeManager(int userId, int removeManagerId, int storeId);
 
@@ -849,11 +862,11 @@ public interface ISession {
      * #XX
      * require XXX?
      *
-     * @param userId the user id
+     * @param userId  the user id
      * @param storeId the id of the store to get info from
-     *               as an owner/founder you can get info about the users working in the store
+     *                as an owner/founder you can get info about the users working in the store
      */
-    List<WorkerCard> getStoreWorkersInfo(int userId, int storeId);
+    Response<List<WorkerCard>> getStoreWorkersInfo(int userId, int storeId);
 
     /**
      * <H1>Conditions</H1>
@@ -875,101 +888,128 @@ public interface ISession {
     Response<Integer> addStorePriceCondition(int storeId, int userId, double lowerBound, double upperBound);
 
     Response<Integer> addStorePriceCondition(int storeId, int userId, double lowerBound);
+
     Response<Integer> addStoreQuantityCondition(int storeId, int userId, int lowerBound, int upperBound);
 
     Response<Integer> addStoreQuantityCondition(int storeId, int userId, int lowerBound);
-    Response<Integer> addCategoryPriceCondition(int storeId, int userId, String category, double lowerBound, double upperBound) ;
+
+    Response<Integer> addCategoryPriceCondition(int storeId, int userId, String category, double lowerBound, double upperBound);
 
 
-    Response<Integer> addCategoryPriceCondition(int storeId, int userId, String category, double lowerBound) ;
+    Response<Integer> addCategoryPriceCondition(int storeId, int userId, String category, double lowerBound);
 
 
-    Response<Integer> addCategoryQuantityCondition(int storeId, int userId, String category, int lowerBound, int upperBound) ;
+    Response<Integer> addCategoryQuantityCondition(int storeId, int userId, String category, int lowerBound, int upperBound);
 
 
-    Response<Integer> addCategoryQuantityCondition(int storeId, int userId, String category, int lowerBound)  ;
+    Response<Integer> addCategoryQuantityCondition(int storeId, int userId, String category, int lowerBound);
 
 
-    Response<Integer> addDateCondition(int storeId, int userId, LocalDateTime lowerBound, LocalDateTime upperBound) ;
+    Response<Integer> addDateCondition(int storeId, int userId, LocalDateTime lowerBound, LocalDateTime upperBound);
 
-    Response<Integer> addDateCondition(int storeId, int userId, LocalDateTime lowerBound) ;
-
-
-    Response<Integer> addProductPriceCondition(int storeId, int userId, int productId, double lowerBound, double upperBound) ;
+    Response<Integer> addDateCondition(int storeId, int userId, LocalDateTime lowerBound);
 
 
-    Response<Integer> addProductPriceCondition(int storeId, int userId, int productId, double lowerBound) ;
+    Response<Integer> addProductPriceCondition(int storeId, int userId, int productId, double lowerBound, double upperBound);
 
 
-    Response<Integer> addProductQuantityCondition(int storeId, int userId, int productId, int lowerBound, int upperBound) ;
-
-    Response<Integer> addProductQuantityCondition(int storeId, int userId, int productId, int lowerBound) ;
-
-    Response<Integer> addTimeCondition(int storeId, int userId, LocalDateTime lowerBound, LocalDateTime upperBound)  ;
+    Response<Integer> addProductPriceCondition(int storeId, int userId, int productId, double lowerBound);
 
 
-    Response<Integer> addTimeCondition(int storeId, int userId, LocalDateTime lowerBound)  ;
+    Response<Integer> addProductQuantityCondition(int storeId, int userId, int productId, int lowerBound, int upperBound);
+
+    Response<Integer> addProductQuantityCondition(int storeId, int userId, int productId, int lowerBound);
+
+    Response<Integer> addTimeCondition(int storeId, int userId, LocalDateTime lowerBound, LocalDateTime upperBound);
 
 
-    Response<Integer> addUserAgeCondition(int storeId, int userId, int lowerBound, int upperBound)  ;
+    Response<Integer> addTimeCondition(int storeId, int userId, LocalDateTime lowerBound);
 
-    Response<Integer> addUserAgeCondition(int storeId, int userId, int lowerBound)  ;
+
+    Response<Integer> addUserAgeCondition(int storeId, int userId, int lowerBound, int upperBound);
+
+    Response<Integer> addUserAgeCondition(int storeId, int userId, int lowerBound);
+
     /**
      * <H1>Discounts</H1>
      */
 
-    Response<Integer> addStoreDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String coupon) ;
+    Response<Integer> addStoreDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String coupon);
 
 
-    Response<Integer> addStoreDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, String coupon)  ;
+    Response<Integer> addStoreDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, String coupon);
 
-    Response<Integer> addStoreDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate) ;
-
-
-    Response<Integer> addStoreDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate)  ;
-
-    Response<Integer> addCategoryDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String category, String coupon) ;
-
-    Response<Integer> addCategoryDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, String category, String coupon) ;
+    Response<Integer> addStoreDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate);
 
 
-    Response<Integer> addCategoryDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String category)  ;
+    Response<Integer> addStoreDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate);
 
-    Response<Integer> addCategoryDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, String category) ;
+    Response<Integer> addCategoryDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String category, String coupon);
 
-
-    Response<Integer> addProductDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, int productId, String coupon) ;
-
-
-    Response<Integer> addProductDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, int productId, String coupon) ;
+    Response<Integer> addCategoryDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, String category, String coupon);
 
 
-    Response<Integer> addProductDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, int productId) ;
+    Response<Integer> addCategoryDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, String category);
+
+    Response<Integer> addCategoryDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, String category);
 
 
-    Response<Integer> addProductDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, int productId)  ;
+    Response<Integer> addProductDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, int productId, String coupon);
 
 
-    Response<List<DiscountInfo>> getStoreDiscounts(int storeId, int userId) ;
+    Response<Integer> addProductDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, int productId, String coupon);
 
 
-    Response<DiscountInfo> getDiscount(int storeId, int userId, int discountId) ;
+    Response<Integer> addProductDiscount(int storeId, int userId, int conditionId, double discountPercentage, LocalDate expirationDate, int productId);
 
 
-    Response<VoidResponse> removeDiscount(int storeId, int userId, int discountId) ;
+    Response<Integer> addProductDiscount(int storeId, int userId, double discountPercentage, LocalDate expirationDate, int productId);
+
+
+    Response<List<DiscountInfo>> getStoreDiscounts(int storeId, int userId);
+
+
+    Response<DiscountInfo> getDiscount(int storeId, int userId, int discountId);
+
+
+    Response<VoidResponse> removeDiscount(int storeId, int userId, int discountId);
 
     /**
      * <H1>Discount Accumulation Tree</H1>
      */
 
-    Response<VoidResponse> addDiscountAsRoot(int storeId, int userId, int discountId) ;
+    Response<VoidResponse> addDiscountAsRoot(int storeId, int userId, int discountId);
 
 
-    Response<VoidResponse> addDiscountToXORRoot(int storeId, int userId, int discountId) ;
+    Response<VoidResponse> addDiscountToXORRoot(int storeId, int userId, int discountId);
 
 
-    Response<VoidResponse> addDiscountToMAXRoot(int storeId, int userId, int discountId)  ;
+    Response<VoidResponse> addDiscountToMAXRoot(int storeId, int userId, int discountId);
 
 
-    Response<VoidResponse> addDiscountToADDRoot(int storeId, int userId, int discountId)  ;
+    Response<VoidResponse> addDiscountToADDRoot(int storeId, int userId, int discountId);
+
+    Response<DiscountAccumulationTreeInfo> getDiscountAccumulationTree(int storeId, int userId);
+
+    Response<VoidResponse> deleteStoreAccumulationTree(int storeId, int userId);
+
+    List<Integer> getStoreOwners(int storeId);
+
+    Response<VoidResponse> addIndividualPermission(int userId, int managerId, int storeId, UserPermissions.IndividualPermission individualPermission);
+
+    Response<VoidResponse> removeIndividualPermission(int userId, int managerId, int storeId, UserPermissions.IndividualPermission individualPermission);
+
+    Response<Integer> getUserIdByUsername(String userName);
+
+    Response<HashMap<Integer, String>> getUserIdsToUsernamesMapper(List<Integer> userIds);
+
+    Response<Boolean> isStoreHidden(int storeId);
+
+    Response<Boolean> isAdmin(int userId);
+
+    Response<Condition> getStorePurchasePolicy(int storeId, int userId);
+
+    Response<VoidResponse> resetStorePurchasePolicy(int storeId, int userId);
+
+    Response<List<StoreInfo>> getAllStores();
 }
