@@ -4,15 +4,19 @@ import BGU.Group13B.backend.Pair;
 import BGU.Group13B.backend.User.Message;
 import BGU.Group13B.backend.User.PurchaseFailedException;
 import BGU.Group13B.backend.User.PurchaseHistory;
+import BGU.Group13B.backend.User.UserPermissions;
 import BGU.Group13B.backend.storePackage.PublicAuctionInfo;
 import BGU.Group13B.backend.System.SystemInfo;
 import BGU.Group13B.backend.storePackage.Review;
 import BGU.Group13B.backend.storePackage.WorkerCard;
+import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.Condition;
+import BGU.Group13B.backend.storePackage.permissions.NoPermissionException;
 import BGU.Group13B.backend.storePackage.newDiscoutns.DiscountInfo;
 import BGU.Group13B.backend.storePackage.newDiscoutns.discountHandler.StoreDiscount;
 import BGU.Group13B.service.entity.ReviewService;
 import BGU.Group13B.service.entity.ServiceBasketProduct;
 import BGU.Group13B.service.entity.ServiceProduct;
+import BGU.Group13B.service.info.DiscountAccumulationTreeInfo;
 import BGU.Group13B.service.info.ProductInfo;
 import BGU.Group13B.service.info.StoreInfo;
 
@@ -38,9 +42,10 @@ public class ProxySession implements ISession {
     }
 
     @Override
-    public void addToCart(int userId, int storeId, int productId) {
+    public Response<VoidResponse> addToCart(int userId, int storeId, int productId) {
         if (realSession != null)
             realSession.addToCart(userId, storeId, productId);
+        return null;
     }
 
     @Override
@@ -101,7 +106,7 @@ public class ProxySession implements ISession {
     }
 
     @Override
-    public List<WorkerCard> getStoreWorkersInfo(int userId, int storeId) {
+    public Response<List<WorkerCard>> getStoreWorkersInfo(int userId, int storeId) {
         if (realSession != null)
             return realSession.getStoreWorkersInfo(userId, storeId);
         return null;
@@ -411,11 +416,120 @@ public class ProxySession implements ISession {
         return Response.success(new VoidResponse());
     }
 
+    @Override
+    public Response<DiscountAccumulationTreeInfo> getDiscountAccumulationTree(int storeId, int userId) {
+        if (realSession != null)
+            return realSession.getDiscountAccumulationTree(storeId, userId);
+
+        return null;
+    }
 
     @Override
-    public void purchaseProposalSubmit(int userId, int storeId, int productId, double proposedPrice, int amount) {
+    public Response<VoidResponse> deleteStoreAccumulationTree(int storeId, int userId) {
+        if (realSession != null)
+            return realSession.deleteStoreAccumulationTree(storeId, userId);
+
+        return Response.success(new VoidResponse());
+    }
+
+    @Override
+    public List<Integer> getStoreOwners(int storeId) {
+        if (realSession != null)
+            return realSession.getStoreOwners(storeId);
+
+        return null;
+    }
+
+    @Override
+    public Response<VoidResponse> addIndividualPermission(int userId, int managerId, int storeId, UserPermissions.IndividualPermission individualPermission) {
+        if (realSession != null)
+            return realSession.addIndividualPermission(userId, managerId, storeId, individualPermission);
+
+        return Response.success(new VoidResponse());
+    }
+
+    @Override
+    public Response<VoidResponse> removeIndividualPermission(int userId, int managerId, int storeId, UserPermissions.IndividualPermission individualPermission) {
+        if (realSession != null)
+            return realSession.removeIndividualPermission(userId, managerId, storeId, individualPermission);
+
+        return Response.success(new VoidResponse());
+    }
+
+    @Override
+    public Response<Integer> getUserIdByUsername(String userName) {
+        if (realSession != null)
+            return realSession.getUserIdByUsername(userName);
+
+        return null;
+    }
+
+    @Override
+    public Response<HashMap<Integer, String>> getUserIdsToUsernamesMapper(List<Integer> userIds) {
+        if (realSession != null)
+            return realSession.getUserIdsToUsernamesMapper(userIds);
+
+        return null;
+    }
+
+    @Override
+    public Response<Boolean> isStoreHidden(int storeId) {
+        if (realSession != null)
+            return realSession.isStoreHidden(storeId);
+
+        return null;
+    }
+
+    @Override
+    public Response<Boolean> isAdmin(int userId) {
+        if (realSession != null)
+            return realSession.isAdmin(userId);
+
+        return null;
+    }
+
+    @Override
+    public Response<Condition> getStorePurchasePolicy(int storeId, int userId) {
+        if (realSession != null)
+            return realSession.getStorePurchasePolicy(storeId, userId);
+
+        return null;
+    }
+
+    @Override
+    public Response<VoidResponse> resetStorePurchasePolicy(int storeId, int userId) {
+        if (realSession != null)
+            return realSession.resetStorePurchasePolicy(storeId, userId);
+
+        return Response.success(new VoidResponse());
+    }
+
+    @Override
+    public Response<List<StoreInfo>> getAllStores() {
+        if (realSession != null)
+            return realSession.getAllStores();
+
+        return null;
+    }
+
+
+    @Override
+    public Response<VoidResponse> purchaseProposalSubmit(int userId, int storeId, int productId, double proposedPrice, int amount) {
         if (realSession != null)
             realSession.purchaseProposalSubmit(userId, storeId, productId, proposedPrice, amount);
+        return Response.success(new VoidResponse());
+    }
+
+    @Override
+    public void purchaseProposalApprove(int managerId, int storeId, int productId) throws NoPermissionException {
+        if(realSession != null)
+            realSession.purchaseProposalApprove(managerId, storeId, productId);
+    }
+
+    @Override
+    public void purchaseProposalReject(int storeId, int managerId, int bidId) throws NoPermissionException {
+        if(realSession != null)
+            realSession.purchaseProposalReject(storeId, managerId, bidId);
     }
 
     @Override
@@ -476,14 +590,14 @@ public class ProxySession implements ISession {
     }
 
     @Override
-    public Response<List<ProductInfo>> filterByPriceRange(int minPrice, int maxPrice) {
+    public Response<List<ProductInfo>> filterByPriceRange(double minPrice, double maxPrice) {
         if (realSession != null)
             realSession.filterByPriceRange(minPrice, maxPrice);
         return null;
     }
 
     @Override
-    public Response<List<ProductInfo>> filterByProductRank(int minRating, int maxRating) {
+    public Response<List<ProductInfo>> filterByProductRank(double minRating, double maxRating) {
         if (realSession != null)
             realSession.filterByProductRank(minRating, maxRating);
         return null;
@@ -497,7 +611,7 @@ public class ProxySession implements ISession {
     }
 
     @Override
-    public Response<List<ProductInfo>> filterByStoreRank(int minRating, int maxRating) {
+    public Response<List<ProductInfo>> filterByStoreRank(double minRating, double maxRating) {
         if (realSession != null)
             realSession.filterByStoreRank(minRating, maxRating);
         return null;
@@ -929,14 +1043,14 @@ public class ProxySession implements ISession {
     }
 
     @Override
-    public Response<String> getUserPurchaseHistory(int userId) {
+    public  Response<List<PurchaseHistory>>  getUserPurchaseHistory(int userId) {
         if (realSession != null)
             return realSession.getUserPurchaseHistory(userId);
         return null;
     }
 
     @Override
-    public Response<String> getUserPurchaseHistoryAsAdmin(int userId, int adminId) {
+    public Response<List<PurchaseHistory>>  getUserPurchaseHistoryAsAdmin(int userId, int adminId) {
         if (realSession != null)
             return realSession.getUserPurchaseHistoryAsAdmin(userId, adminId);
         return null;
